+++ src/ultra/util.clj
@@ -1,0 +1,13 @@
+(ns ultra.util
+  "Handy functions that don't belong anywhere else.")
+
+(defmacro with-err-str
+  "Evaluates exprs in a context in which *err* is bound to a fresh
+  StringWriter.  Returns the string created by any nested printing
+  calls."
+  {:added "0.1.9"}
+  [& body]
+  `(let [s# (new java.io.StringWriter)]
+     (binding [*err* s#]
+       ~@body
+       (str s#))))